Enable custom content-type in `HttpConnection`
