+++ test/tencent_cloud_api_test.dart
@@ -1,20 +1,13 @@
-import 'dart:math';
-
 import 'package:flutter_test/flutter_test.dart';
-import 'package:tencent_cloud_api/HttpConnection.dart';
 
 import 'test_cloud_api.dart';
+import 'test_http_connection.dart';
 import 'test_sign.dart';
 
 void main() {
   group('test sign', testSign);
   
-  test('post request with plaintext', () async {
-    var nextInt = Random().nextInt(1<<16);
-    var response = await HttpConnection().post('https://postman-echo.com/post', data: '$nextInt');
-    var responseBody = response.data;
-    expect(responseBody['data'], '$nextInt');
-  });
+  group('test http connection', testHttpConnection);
 
   test('invoke cloud api', testCloudApi, skip: 'TODO: Read params from environment');
 }
